Create ingestion foundation for roleplaying games.
This commit focuses on enabling developers to process data in whatever source in Rust objects. A number of patterns have been established and fleshed out here. This includes the addition of the two important traits `Ingest` and `RolePlayingGame`. Using these traits as guidelines, you can install data locally as well as recursivly ingest desired data. This diff showcases this with Pathfinder 2nd Edition Classes. Care has been taken with the organization of modules; the goal was to create a comman ingestion pattern for any role playing game.